Migrate all it configs to project
