+++ .github/workflows/populate-it-data.yml
@@ -26,10 +26,8 @@ jobs:
         env:
           CLOUDSQL_SQLSERVER_PASSWORD: ${{ secrets.CLOUDSQL_SQLSERVER_PASSWORD }}
       - name: Temporarily remove conflicting Scala classes # some IT test classes won't compile because they depend on macros based on test data that's missing
-        run: find scio-google-cloud-platform/src/it/scala/com/spotify/scio/bigquery -type f ! -name 'PopulateTestData.scala' -delete
+        run: find integration/src/test/scala/com/spotify/scio/bigquery -type f ! -name 'PopulateTestData.scala' -delete
       - name: Populate BQ test data
-        run: sbt "scio-google-cloud-platform/IntegrationTest/runMain com.spotify.scio.bigquery.PopulateTestData"
+        run: sbt "integration/Test/runMain com.spotify.scio.bigquery.PopulateTestData"
       - name: Populate GCS test data
-        run: sbt "scio-test/IntegrationTest/runMain com.spotify.scio.PopulateTestData"
-      - name: Populate SQL test data
-        run: sbt "scio-jdbc/IntegrationTest/runMain com.spotify.scio.jdbc.PopulateTestData"
+        run: sbt "integration/Test/runMain com.spotify.scio.PopulateTestData"
