+++ integration/src/test/java/org/apache/beam/sdk/extensions/smb/benchmark/CoGroupByKeyBenchmark.java
