+++ integration/src/test/resources/word-count-repl.scala
