+++ integration/src/test/scala/com/spotify/scio/avro/types/AvroTapIT.scala
