+++ integration/src/test/scala/com/spotify/scio/avro/types/AvroTypeIT.scala
