+++ integration/src/test/scala/com/spotify/scio/avro/types/TypeProviderIT.scala
