+++ integration/src/test/scala/com/spotify/scio/bigquery/BigQueryPartitionUtilIT.scala
