+++ integration/src/test/scala/com/spotify/scio/bigquery/TypedBigQueryIT.scala
