+++ integration/src/test/scala/com/spotify/scio/bigquery/dynamic/DynamicBigQueryIT.scala
