--- scio-elasticsearch/es7/src/it/scala/com/spotify/scio/elasticsearch/ElasticsearchIOIT.scala
@@ -1,22 +1,0 @@
-/*
- * Copyright 2023 Spotify AB
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.spotify.scio.elasticsearch
-
-import com.spotify.scio.testing.PipelineSpec
-class ElasticsearchIOIT extends PipelineSpec with ElasticsearchIOBehavior {
-  "Elasticsearch v7" should behave like elasticsearchIO()
-}
